fix: changed price data type for product struct

+++ src/models/types.go
@@ -2,5 +2,5 @@ package models
 
 type Product struct {
 	Name  string `bson:"name" json:"name"`
-	Price uint64 `bson:"price" json:"price"`
+	Price uint32 `bson:"price" json:"price"`
 }
